requirements.txt
streamlit
face-recognition
opencv-python
numpy
pandas
altair
Pillow